pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dj-database-url>=3.1.0",
    "django>=5.2.9",
    "django-bootstrap5>=26.1",
    "django-filter>=25.2",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "ruff>=0.14.10",
]
